N5HTTPFetch: remove request from dataset_exists
The default impl of dataset_exists provided by N5AsyncReader check both
that the directory of the dataset exists and that the dataset attributes
can be read. This is strictly consistent with the behavior of N5Reader.

However, for HTTP this has several downsides:
- It issues 2 requests, rather than 1.
- It requires that GET of the directory path returns 200, which most
  servers do not when directory listing is disabled.
- The dataset attributes are parsed, even though the parsed
  representation is discard.

Instead, N5HTTPReader overrides this impl to only issue an exists check
(200 status) for the dataset attributes path.

Reported by Stephan Gerhard (@unidesigner).

Closes aschampion/n5-wasm#1.